requirements.txt
winrt-runtime
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
winrt-Windows.Media.Control
winrt-Windows.Storage.Streams
